debugs calculate points after battle and impoving deleteActivePlayers()

+++ src/main/java/controller/FinishController.java
@@ -31,11 +31,12 @@ public class FinishController {
         this.battleService = battleService;
     }
 
+
     @GetMapping
     public ModelAndView finishView(HttpServletRequest req,
                                    HttpServletResponse resp) throws IOException {
         User u = userService.getUserAttributeFromSession(req.getSession());
-        Battle b = finishService.calculatePoints(battleService.getBattleById((Integer) req.getSession().getAttribute("battleId")));
+        Battle b = finishService.result(battleService.getBattleById((Integer) req.getSession().getAttribute("battleId")));
         if (u != null && u.equals(battleService.getBattleById(battleService.isUserInBattle2(u.getLogin())).getPlayer1())) {
             return new ModelAndView("finish", "b", b);
         } else if (u != null && u.equals(battleService.getBattleById(battleService.isUserInBattle2(u.getLogin())).getPlayer2())) {
@@ -52,14 +53,21 @@ public class FinishController {
                        HttpServletRequest req) throws IOException {
         User u = userService.getUserAttributeFromSession(req.getSession());
         if (u != null) {
+            if (!finishService.isFirstExit()) {
+                finishService.setFirstExit();
+                resp.sendRedirect("/fs/main");
+                return;
+            }
             Integer battleId = (Integer) req.getSession().getAttribute("battleId");
             Battle b = battleService.getBattleById(battleId);
-            finishService.deleteBattle((Integer) req.getSession().getAttribute("battleId"));
             finishService.deleteActivePlayers(b);
+            finishService.deleteBattle((Integer) req.getSession().getAttribute("battleId"));
+            req.setAttribute("battleId", null);
             resp.sendRedirect("/fs/main/");
         } else {
             resp.sendRedirect("/fs/");
         }
     }
+
 
 }
